1.62.3 Version Change
This changes the version number to 1.62.3 to reflect the bug changes.
It also adds the mod name and my username to clarify what version your playing.
If you are using this as a mod base, please replace this with your own info.

You may have to play with toggling the manager settings for it to show