Organ Top Red Editor Name fix
Changes recording group order so Organ - Top Red will be prioritized over Santa - Bird Light in editor.

Santa and Unused mech have been moved to the bottom of the live editor. The Mouths group has been moved up by the main characters.

I recommend adding some form of exclusion check in the future; Where you could set a certain group as the default to use the names of. Rather than the current system of always using the first name it finds for a bit.
I also recommend adding an option to the Editor to swap to the alternative names for people who may want to use them. Just don't have them as the default. Maybe a drop down option would work the best for it?